Added salt content

+++ components/iq2020/iq2020.cpp
@@ -312,7 +312,7 @@ int IQ2020Component::processIQ2020Command() {
 		if (salt_content != processingBuffer[9]) {
 			salt_content = processingBuffer[9];
 #ifdef USE_SENSOR
-			if (this->salt_content_sensor_) this->salt_content_sensor_->publish_state(processingBuffer[9]);
+			if (this->salt_content_sensor_) this->salt_content_sensor_->publish_state((float)processingBuffer[9]);
 #endif
 		}
 	}
@@ -497,7 +497,7 @@ int IQ2020Component::processIQ2020Command() {
 			// Misc sensors
 			if (this->pcb_temperature_sensor_) this->pcb_temperature_sensor_->publish_state((float)processingBuffer[128]);
 			if (this->salt_power_sensor_ && (salt_power >= 0)) this->salt_power_sensor_->publish_state(salt_power);
-			if (this->salt_content_sensor_ && (salt_content >= 0)) this->salt_content_sensor_->publish_state(salt_content);
+			if (this->salt_content_sensor_ && (salt_content >= 0)) this->salt_content_sensor_->publish_state((float)salt_content);
 #endif
 
 			if (pending_temp != -1) {
